Config and shot sound
Objects redimension to appear in screen and re implemented shots with sound
