+++ GalactaTEC/Assets/Audio.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 799c73d0b4b030f4c91fa73df55a011d
-NativeFormatImporter:
+guid: b91435df9b5544847a38020d56d8223a
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
