+++ GalactaTEC/Assets/Images/Sprites/Weapons/Canvas.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b68261f78cee24049acd39c5e023903d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
