+++ GalactaTEC/Assets/Images/Sprites/Weapons/normalShot.png.meta
@@ -43,7 +43,7 @@ TextureImporter:
   nPOTScale: 0
   lightmap: 0
   compressionQuality: 50
-  spriteMode: 1
+  spriteMode: 2
   spriteExtrude: 8
   spriteMeshType: 1
   alignment: 0
