+++ GalactaTEC/Assets/Images/Sprites/Weapons/normalShotAnim.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: simpleShotAnimation
+  m_Name: normalShotAnim
   serializedVersion: 7
   m_Legacy: 0
   m_Compressed: 0
