+++ GalactaTEC/Assets/Images/Sprites/Weapons/normalShotAnim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e534887d6efe0ff408aa83183d331ce4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
