+++ GalactaTEC/Assets/Scripts/playerController.cs
@@ -1,0 +1,45 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class playerController : MonoBehaviour
+{
+    float maxSpeed = 300f;
+
+    [SerializeField]
+    private GameObject normalShot;
+    
+    [SerializeField]
+    private Transform attack_Point;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        Move();
+        Attack();
+
+    }
+
+    void Move()
+    {
+        Vector3 pos = transform.position;
+        pos.x += Input.GetAxis("Horizontal") * maxSpeed * Time.deltaTime;
+        pos.y += Input.GetAxis("Vertical") * maxSpeed * Time.deltaTime;
+        transform.position = pos;
+    }
+
+    void Attack()
+    {
+
+        if (Input.GetKeyDown(KeyCode.Space))
+        {
+            Instantiate(normalShot, attack_Point.position, Quaternion.identity);
+        }
+    }
+}
