+++ GalactaTEC/UserSettings/EditorUserSettings.asset
@@ -9,25 +9,25 @@ EditorUserSettings:
       value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 5703075007515d08080d0d7643745e44464f1a2e287d7536752b4565e1b76239
-      flags: 0
-    RecentlyUsedSceneGuid-2:
       value: 065057030056085d090f557512210d44134f1c2f2f7f25342f7e1e65b5e13269
       flags: 0
-    RecentlyUsedSceneGuid-3:
+    RecentlyUsedSceneGuid-2:
       value: 00050c5556075f0d0f08087345720e44454f4172787a7e66782a1b62b6b83068
       flags: 0
-    RecentlyUsedSceneGuid-4:
+    RecentlyUsedSceneGuid-3:
       value: 57080555570c0b0c5e0c597444710a4414164a7a2d2c22347d2a1c62e0e3603b
       flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-4:
       value: 55570d550353580954565f7442250f44124f402b287a75337c704d64e7b93260
       flags: 0
-    RecentlyUsedSceneGuid-6:
+    RecentlyUsedSceneGuid-5:
       value: 555057525d055e0e0b565b231275074412161a78752c226729784f37b7b4306d
       flags: 0
-    RecentlyUsedSceneGuid-7:
+    RecentlyUsedSceneGuid-6:
       value: 0108550350535c5a5d0f5d2046730f4445161d297f297569292d1b61b2b9643e
+      flags: 0
+    RecentlyUsedSceneGuid-7:
+      value: 5703075007515d08080d0d7643745e44464f1a2e287d7536752b4565e1b76239
       flags: 0
     RecentlyUsedSceneGuid-8:
       value: 540004525d515a0f5a5a0a7115265b4413154828292b7560797b1965b0b96c68
